Data from the journal Biotropica (pulled from Web of Knowledge)

A dataset containing 10 articles taken from the journal Biotropica. This dataset represents the typical formatted output from references_read() in the refsplitr package. It serves as a testbed for commonly miscategorized names

Usage

BITR

Format

A data frame with 10 rows and 32 variables:

filename

the original filename the text was created from

refID

the unique identifier given to each reference article by references_read()

AB

Abstract

AF

Full Names

AU

Abbreviated names

C1

Addresses

EM

emails

RI

Web of Science ID

OI

OrcID

RP

Reprint Address

TI

Title

UT

Web of Knowledge Unique ID

BP

See url below

CR

See url below

DE

See url below

DI

See url below

EP

See url below

FN

See url below

FU

See url below

PD

See url below

PG

See url below

PT

See url below

PU

See url below

PY

See url below

PM

See url below

SC

See url below

SN

See url below

SO

See url below

TC

See url below

VL

See url below

WC

See url below

Z9

See url below

The remaining codes are described on the Web of Knowledge website: https://images.webofknowledge.com/images/help/WOS/hs_wos_fieldtags.html