Seperates author information in references files from references_read

authors_clean This function takes the output from references_read and cleans the author information.

Usage

authors_clean(references)

Arguments

references

output from references_read

Details

Information on addresses, emails, ORCIDs, etc are matched.

It then attempts to match same author entries together into likely author groups based on common full names, addresses, emails, ORCIDs etc.

Records that are not matched this way have a Jaro-Winkler similiarty analysis metric calculated for all possible matching author names.

This calculates the amount of character similarities based on distance of similar character.

Examples

## Load the refsplitr sample dataset "BITR" 
data(BITR) 
BITR_clean <- authors_clean(BITR)
#> 
#> Splitting author records
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |                                                                      |   0%  |                                                                              |=====================                                                 |  30%  |                                                                              |                                                                      |   0%  |                                                                              |============================                                          |  40%  |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |                                                                      |   0%  |                                                                              |==========================================                            |  60%  |                                                                              |                                                                      |   0%  |                                                                              |=================================================                     |  70%  |                                                                              |                                                                      |   0%  |                                                                              |========================================================              |  80%  |                                                                              |                                                                      |   0%  |                                                                              |===============================================================       |  90%  |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> Splitting addresses.
#> 
#> standardizing country names...
#> 
#> extracting the names of institutions...
#> 
#> extracting cities...
#> 
#> extracting states/provinces...
#> 
#> extracting country...
#> 
#> processing postal codes...
#> 
#> review, correction, and clean-up...
#> 
#> Please be patient - this might take a bit.
#> 
#> (still working on it...)
#> 
#> (getting closer...)
#> 
#> (not much longer...)
#> 
#> (almost done...)
#> 
#> (so close...the end is in sight!)
#> 
#> (this is it - the last step!)
#> 
#> Matching authors
#>   |                                                                              |                                                                      |   0%  |                                                                              |==                                                                    |   2%  |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   4%  |                                                                              |                                                                      |   0%  |                                                                              |=====                                                                 |   7%  |                                                                              |                                                                      |   0%  |                                                                              |======                                                                |   9%  |                                                                              |                                                                      |   0%  |                                                                              |========                                                              |  11%  |                                                                              |                                                                      |   0%  |                                                                              |=========                                                             |  13%  |                                                                              |                                                                      |   0%  |                                                                              |===========                                                           |  15%  |                                                                              |                                                                      |   0%  |                                                                              |============                                                          |  17%  |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |                                                                      |   0%  |                                                                              |===============                                                       |  22%  |                                                                              |                                                                      |   0%  |                                                                              |=================                                                     |  24%  |                                                                              |                                                                      |   0%  |                                                                              |==================                                                    |  26%  |                                                                              |                                                                      |   0%  |                                                                              |====================                                                  |  28%  |                                                                              |                                                                      |   0%  |                                                                              |=====================                                                 |  30%  |                                                                              |                                                                      |   0%  |                                                                              |=======================                                               |  33%  |                                                                              |                                                                      |   0%  |                                                                              |========================                                              |  35%  |                                                                              |                                                                      |   0%  |                                                                              |==========================                                            |  37%  |                                                                              |                                                                      |   0%  |                                                                              |===========================                                           |  39%  |                                                                              |                                                                      |   0%  |                                                                              |=============================                                         |  41%  |                                                                              |                                                                      |   0%  |                                                                              |==============================                                        |  43%  |                                                                              |                                                                      |   0%  |                                                                              |================================                                      |  46%  |                                                                              |                                                                      |   0%  |                                                                              |=================================                                     |  48%  |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |                                                                      |   0%  |                                                                              |=====================================                                 |  52%  |                                                                              |                                                                      |   0%  |                                                                              |======================================                                |  54%  |                                                                              |                                                                      |   0%  |                                                                              |========================================                              |  57%  |                                                                              |                                                                      |   0%  |                                                                              |=========================================                             |  59%  |                                                                              |                                                                      |   0%  |                                                                              |===========================================                           |  61%  |                                                                              |                                                                      |   0%  |                                                                              |============================================                          |  63%  |                                                                              |                                                                      |   0%  |                                                                              |==============================================                        |  65%  |                                                                              |                                                                      |   0%  |                                                                              |===============================================                       |  67%  |                                                                              |                                                                      |   0%  |                                                                              |=================================================                     |  70%  |                                                                              |                                                                      |   0%  |                                                                              |==================================================                    |  72%  |                                                                              |                                                                      |   0%  |                                                                              |====================================================                  |  74%  |                                                                              |                                                                      |   0%  |                                                                              |=====================================================                 |  76%  |                                                                              |                                                                      |   0%  |                                                                              |=======================================================               |  78%  |                                                                              |                                                                      |   0%  |                                                                              |========================================================              |  80%  |                                                                              |                                                                      |   0%  |                                                                              |==========================================================            |  83%  |                                                                              |                                                                      |   0%  |                                                                              |===========================================================           |  85%  |                                                                              |                                                                      |   0%  |                                                                              |=============================================================         |  87%  |                                                                              |                                                                      |   0%  |                                                                              |==============================================================        |  89%  |                                                                              |                                                                      |   0%  |                                                                              |================================================================      |  91%  |                                                                              |                                                                      |   0%  |                                                                              |=================================================================     |  93%  |                                                                              |                                                                      |   0%  |                                                                              |===================================================================   |  96%  |                                                                              |                                                                      |   0%  |                                                                              |====================================================================  |  98%  |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> Pruning groupings...

## The output of authors_clean is a list with two elements, 
## which can be assigend to dataframes.
BITR_review_df <- BITR_clean$review
BITR_prelim_df <- BITR_clean$prelim

## Users can save the these dataframes outside of R as .csv files.
## The "review_df.csv" is then used to review the groupID or authorID 
## assignments and make any necessary corrections. 
## The function "authors_refine" is used to load and merge the changes 
## into R and create a dataframe used for analyses.